fix(NA): completly move transport logic into transport.
